+++ pkg/scheduler/k8s_internal/predicates/config_maps.go
@@ -11,7 +11,6 @@ import (
 	"golang.org/x/exp/maps"
 	v1 "k8s.io/api/core/v1"
 	ksf "k8s.io/kube-scheduler/framework"
-	k8sframework "k8s.io/kubernetes/pkg/scheduler/framework"
 
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/common_info"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/configmap_info"
@@ -50,7 +49,7 @@ func (_ *ConfigMapPredicate) isFilterRequired(_ *v1.Pod) bool {
 }
 
 func (cmp *ConfigMapPredicate) PreFilter(ctx context.Context, _ ksf.CycleState, pod *v1.Pod, _ []ksf.NodeInfo) (
-	*k8sframework.PreFilterResult, *ksf.Status) {
+	*ksf.PreFilterResult, *ksf.Status) {
 	requiredConfigMapNames := getAllRequiredConfigMapNames(pod)
 
 	var missingConfigMaps []string